Make sure we keep META info of zef hosted modules
Should the zef ecosystem ever be offline, we still want to be able
to access the META of a module without having to investigate the
tar file.

As a side effect, in the future we should be able to look at
JSON files added during a period to find out which modules were
added to the ecosystem in that period.