Update META info of archived DB::Model::Easy
There has only ever been one release of this distribution, and it
did not contain any valid version information.  It no longer lives
in any Raku ecosystem apart from the REA, and the repository has been
archived.  Rather than doing an upgrade operation on the repo, it
felt clearer to just make sure that the external meta information of
this distribution is not "*".  The associated tar-file is **not**
altered in any way, and thus still contains a META file with the
version "*".  This should not be a problem should anyone want to
install this distribution now.